Estimation of 2019 Canadian Election results given every voter participated
This paper examines the 2019 Canadian Election result given every voter participated. We are using data collected in the 2017 GSS and 2019 CES to multiple logistic regression models to examine the relationship between age, gender, and education level. This model showed the conservatives would win in a bigger margin in the popular vote over the Liberals if every voter participated in the election.  The relationship found in the analysis is strong although there is one sub-predictor that seems weak. 

This project was produced for course credit at the University of Toronto: STA304 - Surveys, Sampling, and Observational Data.